Version 1.2
Cleans  up better once  a set of ranges  is cleared (the  variables are unlet).
Added  an  option (buffer  or  global  variable called  'consolidateRanges') which,  if  set  to 1,  will  cause  Rangecommand  to  first sort  and  then consolidate  overlapping ranges  into a  single range  before executing  the command. The ranges will remain consolidated once the command is done.